+++ packages/eslint-config/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @sap-cloud-sdk/eslint-config
 
+## 3.0.2
+
 ## 3.0.1
 
 ## 3.0.0