+++ packages/resilience/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @sap-cloud-sdk/resilience
 
+## 3.0.2
+
+### Patch Changes
+
+- @sap-cloud-sdk/util@3.0.2
+
 ## 3.0.1
 
 ### Patch Changes