+++ packages/temporal-de-serializers/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @sap-cloud-sdk/temporal-de-serializers
 
+## 3.0.2
+
+### Patch Changes
+
+- Updated dependencies [2fa8c36a6]
+- Updated dependencies [2fa8c36a6]
+  - @sap-cloud-sdk/odata-common@3.0.2
+
 ## 3.0.1
 
 ### Patch Changes